# Add measure of roundtrip performance for ArrayBuffers in coverall2 test